pyproject.toml
[project]
name = "air-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nest_asyncio",
    "torch",
    "easyocr",
    "torchvision",
    "supervision==0.18.0",
    "openai>=1.60.1",
    "transformers==4.50.3",
    "ultralytics==8.3.70",
    "azure-identity",
    "numpy==1.26.4",
    "opencv-python",
    "opencv-python-headless",
    "gradio",
    "gradio-log",
    "dill",
    "accelerate",
    "timm",
    "einops==0.8.0",
    "paddlepaddle",
    "paddleocr",
    "ruff==0.6.7",
    "pre-commit==3.8.0",
    "pytest==8.3.3",
    "pytest-asyncio==0.23.6",
    "pyautogui==0.9.54",
    "streamlit>=1.38.0",
    "anthropic[bedrock,vertex]>=0.37.1",
    "jsonschema==4.22.0",
    "boto3>=1.28.57",
    "google-auth<3,>=2",
    "screeninfo",
    "uiautomation",
    "dashscope",
    "groq",
    "google-genai",
    "mcp[cli]",
    "setuptools>=80.9.0",
    "pyaudio>=0.2.14",
    "mss>=10.0.0",
    "pyttsx3>=2.98",
    "pyqt5>=5.15.11",
    "mediapipe>=0.10.21",
    "opencv-python>=4.11.0.86",
    "pynput>=1.8.1",
]
